Use the official archive of zlib
zlib.net only has the latest version in the "root". The latest version
is now 1.3, so 1.2.13 is relegated to the 'fossils' archive.

The 'fossils' link is described as: "All released versions of zlib" so
should work in the future too.

I don't see an official SHA256 for the file on the zlib.net site, so
SHA256 calculated as follows:
 - download the tar.gz
 - verify contents manually
 - shasum -a 256 <tarball>