Update gettext version and build libiconv to support build on Sonoma
The build of gettext on Sonoma (at least arm64) ends with:

ld: Undefined symbols:
  _iconv_ostream_create, referenced from:
      <initial-undefines>
clang: error: linker command failed with exit code 1 (use -v to see invocation)

I tried updating gettext to the latest version (which fixes some Sonoma problems
so it is good to do in any case) but this doesn't resolve the problem.

Building libiconv and adding it as a dependency of gettext seems to fix
the issue. I based the autogen args on what is currently used in the corresponding
Homebrew formula.